Configure and publish as npm package
