+++ .npmignore
@@ -74,7 +74,3 @@ typings/
 
 # Serverless directories
 .serverless
-
-
-# End of https://www.gitignore.io/api/node
-dist/
