+++ package.json
@@ -1,19 +1,44 @@
 {
-  "name": "c4",
-  "version": "3.0.0",
-  "description": "Connect Four",
-  "main": "index.html",
-  "repository": "git@github.com:kenrick95/c4.git",
-  "author": "Kenrick <kenrick95@gmail.com>",
+  "name": "@kenrick95/c4",
+  "version": "3.0.5",
+  "description": "Connect Four game",
+  "main": "./dist/lib.js",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/kenrick95/c4.git"
+  },
+  "author": {
+    "name": "Kenrick",
+    "email": "kenrick95@gmail.com",
+    "url": "https://kenrick95.org"
+  },
+  "bugs": {
+    "url": "https://github.com/kenrick95/c4/issues"
+  },
+  "homepage": "https://github.com/kenrick95/c4",
   "license": "MIT",
+  "files": [
+    "src/",
+    "dist/"
+  ],
   "scripts": {
     "start": "parcel index.html",
-    "build": "parcel build index.html",
-    "prettier": "prettier --write \"src/**/*.?(ts|scss|css)\""
+    "build": "yarn clean && parcel build index.html",
+    "build:lib": "yarn clean && parcel build ./src/lib.ts",
+    "prettier": "prettier --write \"src/**/*.?(ts|scss|css)\"",
+    "prepublishOnly": "yarn build:lib",
+    "clean": "shx rm -rf dist/"
   },
+  "keywords": [
+    "c4",
+    "connect-four",
+    "game"
+  ],
   "devDependencies": {
     "parcel-bundler": "^1.10.1",
-    "prettier": "1.14.3"
+    "prettier": "1.14.3",
+    "shx": "^0.3.2",
+    "typescript": "^3.1.1"
   },
   "dependencies": {
     "es6-promise": "^4.2.5"
