+++ src/game/game-local-ai.ts
@@ -1,11 +1,14 @@
-import { Board, BoardPiece } from '../board'
+import { Board, BoardPiece, BoardBase } from '../board'
 import { GameBase } from './game-base'
 import { Player, PlayerHuman, PlayerAi } from '../player'
 import { Utils } from '../utils'
 
 class GameLocalAi extends GameBase {
-  constructor(players: Array<Player>, canvas: HTMLCanvasElement) {
-    super(players, canvas)
+  constructor(players: Array<Player>, board: BoardBase) {
+    super(players, board)
+  }
+  afterMove() {
+    // no-op
   }
 }
 export function initGameLocalAi() {
@@ -14,19 +17,25 @@ export function initGameLocalAi() {
     console.error('Canvas DOM is null')
     return
   }
+  const board = new Board(canvas)
+  const humanPlayer = new PlayerHuman(BoardPiece.PLAYER_1)
   const game = new GameLocalAi(
-    [
-      new PlayerHuman(BoardPiece.PLAYER_1, canvas),
-      new PlayerAi(BoardPiece.PLAYER_2, canvas)
-    ],
-    canvas
+    [humanPlayer, new PlayerAi(BoardPiece.PLAYER_2)],
+    board
   )
+
   game.start()
-  canvas.addEventListener('click', async () => {
+  canvas.addEventListener('click', async (event: MouseEvent) => {
     if (game.isGameWon) {
       game.reset()
       await Utils.animationFrame()
       game.start()
+    } else {
+      const rect = canvas.getBoundingClientRect()
+      const x = event.clientX - rect.left
+      const y = event.clientY - rect.top
+      const column = Utils.getColumnFromCoord({ x: x, y: y })
+      humanPlayer.doAction(column)
     }
   })
 }
