+++ src/game/index.ts
@@ -1,2 +1,3 @@
+export * from './game-base'
 export * from './game-local-2p'
 export * from './game-local-ai'
