+++ src/lib.ts
@@ -1,0 +1,4 @@
+export * from './player'
+export * from './utils'
+export * from './board'
+export * from './game'
